Revert name change from 43bc40 and e03739 back to `edtf`
