+++ edtf/__init__.py
@@ -1,0 +1,6 @@
+from edtf.parser.grammar import parse_edtf
+from edtf.natlang import text_to_edtf
+from edtf.parser.parser_classes import *
+from edtf.convert import dt_to_struct_time, struct_time_to_date, \
+    struct_time_to_datetime, trim_struct_time, struct_time_to_jd, \
+    jd_to_struct_time, old_specs_to_new_specs_expression
