+++ edtf/appsettings.py
