+++ edtf/jdutil.py
