+++ edtf/parser/__init__.py
@@ -1,0 +1,2 @@
+from edtf.parser.grammar import parse_edtf
+from edtf.parser.parser_classes import *
