--- edtf2/__init__.py
@@ -1,6 +1,0 @@
-from edtf2.parser.grammar import parse_edtf
-from edtf2.natlang import text_to_edtf
-from edtf2.parser.parser_classes import *
-from edtf2.convert import dt_to_struct_time, struct_time_to_date, \
-    struct_time_to_datetime, trim_struct_time, struct_time_to_jd, \
-    jd_to_struct_time, old_specs_to_new_specs_expression
