--- edtf2/parser/__init__.py
@@ -1,2 +1,0 @@
-from edtf2.parser.grammar import parse_edtf
-from edtf2.parser.parser_classes import *
